Update 02_Table of Means NSS 70.do
Hi,
Thanks for this. I think we are closer. I worked on the panel B code. See if you can do one more pass using the data in table 1. We need to try and match the mean and median of the individual assets and liabilities line items. I started this for the total, see if you can check the individual line items and see which ones are the most off.